Basic OAuth flow
This is the minimum required to send a user through the Singly OAuth flow for
any given service. I want to use something like OmniAuth, but that will take
some plugins and maybe a bit of magic to handle different service parameters, so
it'll come later. For now, someone should be able to look at this and understand
what they need to do.